Add the 'task' convenience method to BaseTask, so that Tasks as well as commands (RoboFiles) may use $this->task('Foo', $a, $b).
