+++ src/Task/Archive/Extract.php
@@ -94,10 +94,10 @@ class Extract extends BaseTask
                 rename($extractLocation, $this->to);
             }
             if ($hasEncapsulatingFolder && !$this->preserveTopDirectory) {
-                $result = $this->getContainer()->get('taskFilesystemStack')->rename($filesInExtractLocation[0], $this->to)->run();
+                $result = $this->task('FilesystemStack')->rename($filesInExtractLocation[0], $this->to)->run();
                 @rmdir($extractLocation);
             } else {
-                $result = $this->getContainer()->get('taskFilesystemStack')->rename($extractLocation, $this->to)->run();
+                $result = $this->task('FilesystemStack')->rename($extractLocation, $this->to)->run();
             }
         }
         $this->stopTimer();
