+++ src/Task/BaseTask.php
@@ -3,6 +3,7 @@ namespace Robo\Task;
 
 use Robo\Common\Configuration;
 use Robo\Common\TaskIO;
+use Robo\TaskAccessor;
 use Robo\Collection\Collectable;
 use Robo\Contract\TaskInterface;
 use Psr\Log\LoggerAwareInterface;
@@ -13,6 +14,8 @@ use League\Container\ContainerAwareTrait;
 abstract class BaseTask implements TaskInterface, LoggerAwareInterface, ContainerAwareInterface
 {
     use TaskIO; // uses LoggerAwareTrait
+    use TaskAccessor;
+
     use ContainerAwareTrait;
     use Configuration;
     use Collectable;
